Exceptions generalized and ChatParticipant delete logic updated. #60

+++ src/main/java/com/gamerum/backend/usecase/service/chat/impl/ChatParticipantServiceImpl.java
@@ -7,8 +7,10 @@ import com.gamerum.backend.external.persistence.relational.entity.Profile;
 import com.gamerum.backend.external.persistence.relational.repository.ChatParticipantRepository;
 import com.gamerum.backend.external.persistence.relational.repository.ChatRepository;
 import com.gamerum.backend.external.persistence.relational.repository.ProfileRepository;
-import com.gamerum.backend.usecase.exception.ChatParticipantExistsException;
+import com.gamerum.backend.usecase.exception.AlreadyParticipatedException;
+import com.gamerum.backend.usecase.exception.NotAllowedException;
 import com.gamerum.backend.usecase.exception.NotFoundException;
+import com.gamerum.backend.usecase.exception.NotParticipatedException;
 import com.gamerum.backend.usecase.service.chat.ChatParticipantService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -28,8 +30,7 @@ public class ChatParticipantServiceImpl implements ChatParticipantService {
     private ProfileRepository profileRepository;
 
     @Override
-    public ChatParticipant createChatParticipant(long chatId, ChatParticipantCreateDTO chatParticipantCreateDTO)
-            throws ChatParticipantExistsException {
+    public ChatParticipant createChatParticipant(long chatId, ChatParticipantCreateDTO chatParticipantCreateDTO) {
 
         Chat chat = chatRepository.findById(chatId).orElseThrow(() ->
                 new NotFoundException(Chat.class, "ID", chatId));
@@ -38,14 +39,17 @@ public class ChatParticipantServiceImpl implements ChatParticipantService {
                 .orElseThrow(() -> new NotFoundException(Profile.class, "ID", chatParticipantCreateDTO.getProfileId()));
 
         if (chatParticipantRepository.existsByChatIdAndProfileId(chatId, profile.getId()))
-            throw new ChatParticipantExistsException(profile.getNickname());
+            throw new AlreadyParticipatedException(profile.getNickname());
 
         return chatParticipantRepository.save(
                 new ChatParticipant(profile, chat, chatParticipantCreateDTO.getAddedByProfileId(), false));
     }
 
     @Override
-    public void deleteByIdChatParticipant(Long chatParticipantId) {
+    public void deleteByIdChatParticipant(Long chatId, Long chatParticipantId, Long deleterProfileId){
+        ChatParticipant participant = chatParticipantRepository.findByChatIdAndProfileId(chatId, deleterProfileId)
+                        .orElseThrow(NotParticipatedException::new);
+        if (!participant.isAdmin()) throw new NotAllowedException();
         chatParticipantRepository.deleteById(chatParticipantId);
     }
